Don't shell out to check if user is in group
Summary:
Originally I thought the best way to check if the squasfs mount directory was owned by a group the user was in was to shell out to the `id` command on macOS.

It turns out that (unsurprisingly) [Apple cheats in the `id` command](https://opensource.apple.com/source/shell_cmds/shell_cmds-116/id/id.c).  They use an undocumented syscall, `getgrouplist_2` which sidesteps the issue where `getgrouplist` may not return *all* the groups a user is a member of.

You need to manually specify the function prototype and give it C linkage, but it works. This allows us to not have to shellout on macOS hosts and just query the system (i.e. OpenDirectory) directly.

Reviewed By: chipturner

Differential Revision: D7178770

fbshipit-source-id: 19293603ac330e05d1cf5f517e06aa0e3f7e5928