Don't fail on chown fail
Summary:
This was kind of insidious and took a lot of debugging. Originally a call to `chown` was added after `mkdir` because on darwin `mkdir` defaults the new directory group to that of the enclosing directory. For non-root users this was a problem - they would not be able to execute xars because the xar mountpoint was not group-owned by a group they were a member of.

However, the `PCHECK_SIMPLE` around the return code of the `chown` of the actual mount path is bad because that `chown` consistently fails after the first invocation, likely because the directory is already `chown`d to the propery uid/gid and osxfuse is weird about uids/gids.

Conditionally `chown`ing the directory only if it was created fixes this issue.

Reviewed By: chipturner

Differential Revision: D7283654

fbshipit-source-id: 298318e7e8abcbf370541a265e5a62ca876389cf