Add bdist_xar distutils command
Summary:
Adds a `bdist_xar` distutils command that can construct XARs, so long as
the dependencies are installed. There are no known limitations, other than
downloading dependencies, but I'm sure there are unknown limitations.

We build a wheel, and add it to the XAR.
We find all the common dependencies and add them to the XAR.
We can add both zipped, and installed wheels.
For each console script, optionally specified by `--console-scripts`, we
add any extra dependencies, and build a XAR.

Wishlist for future features:
* Allow installation of dependencies instead of failing.

Reviewed By: cooperlees

Differential Revision: D7230946

fbshipit-source-id: e49ba537e900213c75d3960b336f60484ae75c4c