Improve bdist_xar on OS X (#5)
Summary:
The install paths on OS X with Python installed with Homebrew are not accurate. The `site-packages` is a symlink pointing to the right place, but the scripts/headers install locations are wrong. If we fail to find the package in `src_paths`, use heuristics to guess the kind and prefix. This shouldn't affect any platforms that were already working, since it only happens if normal method fails.
Closes https://github.com/facebookincubator/xar/pull/5

Reviewed By: cooperlees

Differential Revision: D8683469

Pulled By: terrelln

fbshipit-source-id: 4acf6c1779bb002820d7631de46c99a81514ad6c