Download using PIP
Summary:
`pkg_resources.working_set.resolve()` takes an `installer` callback which is used to download a requirement if it isn't present. Add an optional (disabled with `--no-download`) installer to `bdist_xar` that uses pip to download the dependencies.
* pip is located using `pkg_resources` to make sure we get the right pip for the Python environment.
* We call the pip entry point in a separate process to download (but not install) the requirement and its dependencies.
* We add some pkg_resources finders to make `pkg_resources.find_distributions()` work with wheel archives.
* Extract and build every source distribution that we installed. We call the distributions `setup.py` with small modifications to build the wheel.
* Add each wheel to the `working_set` so we can find it later, and return the newly downloaded wheel distribution.

Since we use the `pip` installed in the running (virtual) environment, we pick up the pip config automatically. For example we automatically download from Facebook's PyPI instance when in a `pyenv`.

Reviewed By: cooperlees

Differential Revision: D8776579

fbshipit-source-id: 09c83b3081111b408757edd579ef79cb05c1d258