Allow duplicate files when sha256 matches
Summary:
Before if any two packages include the same file, or if the same package
includes a file twice, we would fail. Now, when we are copying an
installation allow duplicates when `force=False` if the sha256 matches.

`WheelFile.install()` doesn't have a similar option. The next version of
wheel removes `WheelFile`, so when I fix that, I will make the same change
for installing from an archive. Until then, just set `force=True`.

Fixes #13.

Reviewed By: cooperlees

Differential Revision: D9011530

fbshipit-source-id: c70aba20ff0a8a56f8c4be23d18c7f5ce79b87f4