replace xarexec_timeout_killswitch with more general override
Summary:
There was an existing test for a sentinel file (xarexec_timeout_killswitch),
which could be used to disable timeout processing in squashfuse.

This is replaced with a more generic mechanism which can specify a timeout
value, where zero is disabling. The intent is to allow experimentation
with different timeout settings.

Reviewed By: chipturner

Differential Revision: D13824343

fbshipit-source-id: afba2e55b32223d971edee569a6179bd0539d114