Vendor in wheel
Summary:
Upstream wheel refactored and deleted an API we depend on.
We would either have to copy in the specific code we need,
and potentially make modifications, or vendor in wheel. I
chose to vendor in wheel because it is cleaner, and we can
go back to upstream if they provide the API we need again.

See https://github.com/pypa/wheel/issues/262.

Reviewed By: cooperlees

Differential Revision: D14146161

fbshipit-source-id: 5f6a3e49ce5c6d19964a3b067ec79ca6985d0d33