Use the PID namespace as the default mount seed
Summary:
Before this diff, if a XAR is run from multiple pid namespaces, they will share a mount point (same mount namespace), backed by one process, which is in one of the pid namespaces.  When one of those pid namespaces exits, it may take the `squashfuse_ll` process with it, destroying the mount point *which may be still actively used by the other running XAR*.  This causes sporadic crashes in XARs.

We most commonly see this in systemd timers, where we try to use a seed via chef rules, but this change will be more thorough and handle timers, containers, etc not handled by our custom chef rules.  This is causing an increase in errors after the stack of D22164721 landed, and may be contributing to other sporadic XAR exits in other environments (e.g. edgeOS).

Note the format here is compatible with seeds (i.e. not adding a separate field in addition to the seed).  This is to avoid issues where something may parse the mountpoint (which is bad, but who knows what people do).  Another option though would be to always add it.  Finally, we could combine the pid and mount namespace (say, with a hash) but that makes it impossible to simply see the numeric namespace ID which can be useful debugging.

Reviewed By: terrelln

Differential Revision: D22494175

fbshipit-source-id: 276098f847c2d7d9c1cbb90c3a2ab34acf711064