Avoid a race condition on startup by opening the persistent fd sooner
Summary:
Before this diff, there is a brief chance of a race condition in XARs.  The statfs call could succeed but then an idle unmount could sneak between that and actually using files inside the mount.  This is because the statfs call doesn't (and can't) count as an access against idle timeouts.

This change moves the opening of `bootstrap_fd` earlier.  If the open succeeds, we just keep that fd.  If it fails, we later will retry after mounting has occurred.  Note that this has a slight change in that `-m` opens the file before exiting.  This should be fine and actually makes `-m` less racy anyway.

Reviewed By: terrelln

Differential Revision: D22530138

fbshipit-source-id: 42dcc60cba579ed97cfb8b02c32dfb8a3dc017b2