Add unxar executable
Summary:
Add `unxar` executable that extracts the contents of a XAR to a directory using `unsquashfs`.
This is needed because we need to read the `"OFFSET"` header to see the offset. It is currently
only 4096, but that could potentially change in the future, so it is best to use a tool that
we know about, instead of accumulating magic numbers across the codebase.

Reviewed By: chipturner

Differential Revision: D23455435

fbshipit-source-id: 80fa3ec0ee96b003c87d029a4d96db4f460d98af