Support running XARs out of the expanded filesystem without xarexec's help
Summary:
This diff creates a new file in the root of the squashfs filesystem that is a symlink from the name of the build target (e.g. `xar_subsystem_check`) to `xar_bootstrap.sh`.  The bootstrap now can detect if it is invoked as a symlink and run directly from the filesystem just as if it were invoked via `xarexec`.

It also provides `invoke_xar_via_trampoline` which is a fixed name that executes the XAR for cases where knowing the original build target is inconvenient or impractical.

This allows for packaging a XAR as raw bytes on disk (say, inside an fbpkg or as a kernel squashfs mount) and running a usefully named executable that will act as if it were simply the XAR.

Reviewed By: terrelln

Differential Revision: D26878713

fbshipit-source-id: a11029c03a41e273c80a2aa05a834610a95c647e