Support running XARs out of the expanded filesystem without xarexec's help
Summary:
This diff creates a new file in the root of the squashfs filesystem that is a symlink from the name of the build target (e.g. `xar_subsystem_check`) to `xar_bootstrap.sh`.  The bootstrap now can detect if it is invoked as a symlink and run directly from the filesystem just as if it were invoked via `xarexec`.

It also provides `invoke_xar_via_trampoline` which is a fixed name that executes the XAR for cases where knowing the original build target is inconvenient or impractical.

This allows for packaging a XAR as raw bytes on disk (say, inside an fbpkg or as a kernel squashfs mount) and running a usefully named executable that will act as if it were simply the XAR.

This is a retry of D26878713 (https://github.com/facebookincubator/xar/commit/264ec10e04265360d5a3ebdecd8c33be0c1c4236) which broke when landed (https://fb.workplace.com/groups/btrmeup/permalink/3855081081238196/).  The only difference is escaping `=` with `\x3d` which is shell-compatible and doesn't trigger the xarexec bug.

Reviewed By: terrelln

Differential Revision: D27057029

fbshipit-source-id: 6450f4d92fd83ac89df9087f9901769b36b202f0