Verify XAR output at build time and log to Scuba on failure
Summary:
Verify XAR output at par build time. Any XAR that is created should
strictly adhere to the contract. This will reduce the chance of
failure when attempting to parse/mount the XAR at runtime (for
example, in TW agent).

We do not enforce yet but instead log to scuba, to be sure everything
is safe.

https://www.internalfb.com/intern/wiki/XAR/XAR_Header/

Reviewed By: snarkmaster

Differential Revision: D29701904

fbshipit-source-id: 061f37c3ff2775d775240c2dfaeb131caf8cc4f7