improve switch statement handling in XarParser.cpp
Summary:
D30926949 (https://github.com/facebookincubator/xar/commit/c66dc48572649707a023a98223c1c830b5acfbbd) fixed a compile error in XarParser.cpp about not returning a value
in all cases.  However, it fixed the problem by adding a `default` case to a
switch statement switching on enum types.  Adding a `default` case is
undesirable here, as it will prevent the compiler from warning if new enum
types are added in the future without updating this switch statement.

This fixes the problem by moving the `return` statement outside of the switch,
so that the switch statement does not have a default case, but the `return`
statement will be hit if it falls through.

Reviewed By: alexsn

Differential Revision: D31212626

fbshipit-source-id: 0c49af1c2089540d38bb6e5008c66b77a03678f9