tools/xar/XarParser.cpp: fix platform010-exposed build errors
Summary:
We should all be migrating to platform010, for the improved performance of its
generated code and for its improved diagnostics/portability.

That diagnostic was rather opaque, but resolves to "use const in the lambda, or else nothing matches".

This avoids the following errors:

  third-party-buck/platform010/build/libgcc/include/c++/trunk/bits/stl_numeric.h:169:11: error: no matching function for call to object of type '(lambda at tools/xar/XarParser.cpp:333:9)'

# Context:
The cost of delaying a platform upgrade is now estimated at about third-party-buck/platform010/build/libgcc/include/c\+\+/trunk/bits/stl_numeric.hM/week.
For the preceding one, it was measured at [**That diagnostic was rather opaque, but resolves to "use const in the lambda, or else nothing matches".M/week**](https://fburl.com/jfmpsqmy).

Differential Revision: D32322020

fbshipit-source-id: a28770af7c5edab8691a3127a6031fc3772c164d