Support building XAR packages on fedora using `make_xar`
Summary:
**Summary**
Check for fedora as well as centos in `make_xar` to allow it's use on a managed linux machine.

**Background**
We are using `make_xar` as a dotslash binary as part of a buck build in `arvr`. This works well on a devserver or ondemand but fails on a managed linux desktop with the error below.

`make_xar.py` explicitly check for centos using `/etc/centos-release` when setting the group id for the user `nobody`, using `nobody` for centos and `nogroup` otherwise. Like centos, however, fedora expects the group to be `nobody` (see [standard users](https://docs.fedoraproject.org/en-US/Fedora/14/html/Deployment_Guide/s1-users-groups-standard-groups.html) and [standard groups](https://docs.fedoraproject.org/en-US/Fedora/14/html/Deployment_Guide/s1-users-groups-standard-users.html)), leading to this failure

**This diff**
 Here, we instead check for `/etc/redhat-release` which is present on both fedora and centos.

Reviewed By: pixelb

Differential Revision: D34074349

fbshipit-source-id: fcec3e018c777700622b4a2a54ba70381feda9e1