Fix `xar_subsystem_check` (and thereby `tools/xar:xar_subsystem_unittest`) to be NativePython compatible
Summary:
It seems enabling NativePython, which moved the shared
library code from `libomnibus.so`, also removed `libcxx-build-info.so`
from the build.  This broke a check in `xar_subsystem_check` which
then broke the unit test.

This change removes the expectation `libcxx-build-info.so` be present
in the XAR but still enforces at least _one_ file in the XAR is
mmap'd, which could be a shared library or the python interpreter
itself (which, at the time of this writing, is
`$FB_PAR_RUNTIME_FILES/runtime/bin/xar_subsystem_check#native-main#platform-runtime#python#py_version_3_10`)

Reviewed By: zsol

Differential Revision: D50127111

fbshipit-source-id: bae9949e4a8588698d574bda9977a7658fe40d86