Fix dependencies to allow builds in 2023 (#55)
Summary:
Right out of the box `cargo build` don't work for `clean_xar_mounts` in 2023.
It turns out that something is wrong with the 2.1.x version of slog, I didn't look very hard to know what. It seems that upgrading to something a bit newer is working.
Similarly using 1.0 version of anyhow is not working (complains about missing backtrace) instead using a recent version and enabling the `backtrace` feature is solving the problem

Pull Request resolved: https://github.com/facebookincubator/xar/pull/55

Reviewed By: dtolnay

Differential Revision: D52439947

Pulled By: itamaro

fbshipit-source-id: 1a6314361daea777cdae08383f26a0f692a97671